+++ src/lib/isCurrency.js
@@ -2,6 +2,8 @@ import merge from './util/merge';
 import assertString from './util/assertString';
 
 function currencyRegex(options) {
+  let decimal_digits = `\\d{${options.digits_after_decimal[0]}}`;
+  options.digits_after_decimal.forEach((digit, index) => { if (index !== 0) decimal_digits = `${decimal_digits}|\\d{${digit}}`; });
   const symbol =
     `(\\${options.symbol.replace(/\./g, '\\.')})${(options.require_symbol ? '' : '?')}`,
     negative = '-?',
@@ -10,8 +12,8 @@ function currencyRegex(options) {
     valid_whole_dollar_amounts = [
       '0', whole_dollar_amount_without_sep, whole_dollar_amount_with_sep],
     whole_dollar_amount = `(${valid_whole_dollar_amounts.join('|')})?`,
-    decimal_amount = `(\\${options.decimal_separator}\\d{2})?`;
-  let pattern = whole_dollar_amount + decimal_amount;
+    decimal_amount = `(\\${options.decimal_separator}(${decimal_digits}))${options.require_decimal ? '' : '?'}`;
+  let pattern = whole_dollar_amount + (options.allow_decimal || options.require_decimal ? decimal_amount : '');
 
   // default is negative sign before symbol, but there are two other options (besides parens)
   if (options.allow_negatives && !options.parens_for_negatives) {
@@ -63,6 +65,9 @@ const default_currency_options = {
   allow_negative_sign_placeholder: false,
   thousands_separator: ',',
   decimal_separator: '.',
+  allow_decimal: true,
+  require_decimal: false,
+  digits_after_decimal: [2],
   allow_space_after_digits: false,
 };